pyproject.toml
[project]
name = "ascii-motion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
]
